% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackageSp21-package.R
\docType{package}
\name{testpackageSp21-package}
\alias{testpackageSp21}
\alias{testpackageSp21-package}
\title{testpackageSp21: My Package as a Demo Package}
\description{
This is a demo package.
}
\author{
\strong{Maintainer}: Madeline Beasley \email{mbeasley315@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
